+++ src/main/java/kr/co/easystock/domain/inquiry/InquiryRepository.java
@@ -3,6 +3,6 @@ package kr.co.easystock.domain.inquiry;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface InquiryRepository extends JpaRepository<Inquiry, Integer>, JpaSpecificationExecutor<Inquiry>
+public interface InquiryRepository extends JpaRepository<Inquiry, Long>, JpaSpecificationExecutor<Inquiry>
 {
 }
